Move PreHeader to ergotree-mir
