+++ ergo-lib/src/chain/ergo_state_context.rs
@@ -1,4 +1,5 @@
 //! Blockchain state
+use ergotree_ir::mir::header::PreHeader;
 
 /// Blockchain state (last headers, etc.)
 #[derive(PartialEq, Eq, Debug, Clone)]
@@ -14,20 +15,5 @@ impl ErgoStateContext {
         ErgoStateContext {
             pre_header: PreHeader::dummy(),
         }
-    }
-}
-
-/// Block header with the current `spendingTransaction`, that can be predicted
-/// by a miner before it's formation
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub struct PreHeader {
-    /// Block height
-    pub height: i32,
-}
-
-impl PreHeader {
-    /// Dummy instance intended for tests where actual values are not used
-    pub fn dummy() -> Self {
-        PreHeader { height: 0 }
     }
 }
